Rename UpdateAccountCommand[Handler] => UpdateAccountNameCommand[Handler]
