+++ src/SFA.DAS.EmployerFinance.MessageHandlers/EventHandlers/EmployerAccounts/ChangedAccountNameEventHandler.cs
@@ -2,7 +2,7 @@ using System.Threading.Tasks;
 using MediatR;
 using NServiceBus;
 using SFA.DAS.EmployerAccounts.Messages.Events;
-using SFA.DAS.EmployerFinance.Application.Commands.UpdateAccount;
+using SFA.DAS.EmployerFinance.Application.Commands.UpdateAccountName;
 
 namespace SFA.DAS.EmployerFinance.MessageHandlers.EventHandlers.EmployerAccounts
 {
@@ -17,7 +17,7 @@ namespace SFA.DAS.EmployerFinance.MessageHandlers.EventHandlers.EmployerAccounts
 
         public Task Handle(ChangedAccountNameEvent message, IMessageHandlerContext context)
         {
-            return _mediator.Send(new UpdateAccountCommand(message.AccountId, message.CurrentName, message.Created));
+            return _mediator.Send(new UpdateAccountNameCommand(message.AccountId, message.CurrentName, message.Created));
         }
     }
 }
