+++ src/SFA.DAS.EmployerFinance.UnitTests/MessageHandlers/EmployerAccounts/ChangedAccountNameEventHandlerTests.cs
@@ -1,7 +1,7 @@
 using System.Threading.Tasks;
 using NUnit.Framework;
 using SFA.DAS.EmployerAccounts.Messages.Events;
-using SFA.DAS.EmployerFinance.Application.Commands.UpdateAccount;
+using SFA.DAS.EmployerFinance.Application.Commands.UpdateAccountName;
 using SFA.DAS.EmployerFinance.MessageHandlers.EventHandlers.EmployerAccounts;
 using SFA.DAS.Testing;
 
@@ -14,7 +14,7 @@ namespace SFA.DAS.EmployerFinance.UnitTests.MessageHandlers.EmployerAccounts
         [Test]
         public Task Handle_WhenHandlingChangedAccountNameEvent_ThenShouldSendUpdateAccountNameCommand()
         {
-            return TestAsync(f => f.Handle(), f => f.VerifySend<UpdateAccountCommand>((c, m) => 
+            return TestAsync(f => f.Handle(), f => f.VerifySend<UpdateAccountNameCommand>((c, m) => 
                 c.AccountId == m.AccountId && c.Name == m.CurrentName && c.Updated == m.Created));
         }
     }
