+++ src/SFA.DAS.EmployerFinance/Application/Commands/UpdateAccountName/UpdateAccountNameCommandHandler.cs
@@ -5,18 +5,18 @@ using MediatR;
 using Microsoft.EntityFrameworkCore;
 using SFA.DAS.EmployerFinance.Data;
 
-namespace SFA.DAS.EmployerFinance.Application.Commands.UpdateAccount
+namespace SFA.DAS.EmployerFinance.Application.Commands.UpdateAccountName
 {
-    public class UpdateAccountCommandHandler : AsyncRequestHandler<UpdateAccountCommand>
+    public class UpdateAccountNameCommandHandler : AsyncRequestHandler<UpdateAccountNameCommand>
     {
         private readonly Lazy<EmployerFinanceDbContext> _db;
 
-        public UpdateAccountCommandHandler(Lazy<EmployerFinanceDbContext> db)
+        public UpdateAccountNameCommandHandler(Lazy<EmployerFinanceDbContext> db)
         {
             _db = db;
         }
 
-        protected override async Task Handle(UpdateAccountCommand request, CancellationToken cancellationToken)
+        protected override async Task Handle(UpdateAccountNameCommand request, CancellationToken cancellationToken)
         {
             var account = await _db.Value.Accounts.SingleAsync(a => a.Id == request.AccountId, cancellationToken);
